+++ src/Genius/Authentication/OAuth2.php
@@ -15,8 +15,7 @@ use Psr\Http\Message\RequestInterface;
  */
 final class OAuth2 implements Authentication
 {
-    
-    const API_URL = 'https://api.genius.com/oauth/';
+    protected const API_URL = 'https://api.genius.com/oauth/';
     
     /**
      * @var string
@@ -65,7 +64,7 @@ final class OAuth2 implements Authentication
      * @param Scope $scope
      * @param HttpClient $httpClient
      */
-    public function __construct($client_id, $client_secret, $redirect_uri, Scope $scope, HttpClient $httpClient = null)
+    public function __construct(string $client_id, string $client_secret, string $redirect_uri, Scope $scope, ?HttpClient $httpClient = null)
     {
         $this->clientId = $client_id;
         $this->clientSecret = $client_secret;
@@ -78,63 +77,62 @@ final class OAuth2 implements Authentication
         }
         $this->messageFactory = MessageFactoryDiscovery::find();
         $this->setScope($scope);
-        $this->state = $this->getRandomState();
+        $this->setState($this->getRandomState());
     }
-    
-    public function setScope(Scope $scope)
+
+    public function setScope(Scope $scope): OAuth2
     {
         $this->scope = $scope;
         
         return $this;
     }
-    
-    public function setClientId($clientId)
+
+    public function setClientId(string $clientId): OAuth2
     {
         $this->clientId = $clientId;
         
         return $this;
     }
-    
-    public function setAccessToken($acces_token)
+
+    public function setAccessToken(string $access_token): OAuth2
     {
-        $this->accessToken = $acces_token;
-        
+        $this->accessToken = $access_token;
+
         return $this;
     }
-    
-    public function setRedirectUri($redirect_uri)
+
+    public function setRedirectUri(string $redirect_uri): OAuth2
     {
         $this->redirectUri = $redirect_uri;
         
         return $this;
     }
-    
-    public function setMessageFactory(MessageFactory $messageFactory)
+
+    public function setMessageFactory(MessageFactory $messageFactory): OAuth2
     {
         $this->messageFactory = $messageFactory;
         
         return $this;
     }
-    
-    public function setState($state)
+
+    public function setState(string $state): OAuth2
     {
         $this->state = $state;
         
         return $this;
     }
-    
-    public function getAccessToken()
+
+    public function getAccessToken(): ?string
     {
         if ($this->hasValidAccessToken()) {
             return $this->accessToken;
         }
         
         $this->accessToken = null;
-        
-        return false;
+        return null;
     }
-    
-    protected function getHttpClient()
+
+    protected function getHttpClient(): HttpClient
     {
         if ($this->httpClient === null) {
             $this->httpClient = HttpClientDiscovery::find();
@@ -142,37 +140,30 @@ final class OAuth2 implements Authentication
         
         return $this->httpClient;
     }
-    
-    public function getAuthUrl()
+
+    public function getAuthUrl(): string
     {
         return self::API_URL . 'authorize?client_id=' . $this->clientId .
             '&redirect_uri=' . $this->redirectUri . '&scope=' . $this->scope . '&state=' . 'randstate' .
             '&response_type=code';
     }
-    
-    public function getState()
+
+    public function getState(): string
     {
         return $this->state;
     }
-    
-    /**
-     * @return Scope
-     */
-    public function getScope()
+
+    public function getScope(): Scope
     {
         return $this->scope;
     }
     
-    public function hasValidAccessToken()
+    public function hasValidAccessToken(): bool
     {
-        if ($this->accessToken !== null) {
-            return true;
-        }
-        
-        return false;
+        return $this->accessToken !== null;
     }
     
-    public function refreshToken($code)
+    public function refreshToken(string $code): ?string
     {
         if ($this->getAccessToken()) {
             return $this->getAccessToken();
@@ -180,7 +171,10 @@ final class OAuth2 implements Authentication
         
         $request = $this->httpClient->sendRequest(
             $this->messageFactory->createRequest(
-                'POST', self::API_URL . 'token', [], http_build_query([
+                'POST',
+                self::API_URL . 'token',
+                [],
+                http_build_query([
                     'code' => $code,
                     'client_secret' => $this->clientSecret,
                     'grant_type' => 'authorization_code',
@@ -192,17 +186,16 @@ final class OAuth2 implements Authentication
         );
         
         if ($request->getStatusCode() === 200) {
-            $body = json_decode($request->getBody());
-            
+            $body = json_decode($request->getBody()->getContents(), false);
+
             $this->accessToken = $body->access_token;
-            
             return $this->accessToken;
         }
-        
-        return false;
+
+        return null;
     }
-    
-    protected function getRandomState($length = 32)
+
+    protected function getRandomState(int $length = 32): string
     {
         // Converting bytes to hex will always double length. Hence, we can reduce
         // the amount of bytes by half to produce the correct length.
@@ -212,7 +205,7 @@ final class OAuth2 implements Authentication
     /**
      * {@inheritdoc}
      */
-    public function authenticate(RequestInterface $request)
+    public function authenticate(RequestInterface $request): RequestInterface
     {
         if ($this->hasValidAccessToken()) {
             $header = sprintf('Bearer %s', $this->accessToken);