+++ src/Genius/Resources/SearchResource.php
@@ -9,10 +9,8 @@ namespace Genius\Resources;
  */
 class SearchResource extends AbstractResource
 {
-    
-    public function get($query)
+    public function get(string $query): \stdClass
     {
         return $this->sendRequest('GET', 'search/?' . http_build_query(['q' => $query]));
     }
-    
 }