+++ src/Genius/Resources/SongsResource.php
@@ -10,10 +10,8 @@ namespace Genius\Resources;
  */
 class SongsResource extends AbstractResource
 {
-    
-    public function get($id, $text_format = 'dom')
+    public function get(int $id, string $text_format = 'dom'): \stdClass
     {
         return $this->sendRequest('GET', 'songs/' . $id . '/?' . http_build_query(['text_format' => $text_format]));
     }
-    
 }